Brands with hybrid runway, allow routing
