+++ config/rapidez/statamic.php
@@ -19,10 +19,6 @@ return [
 
     // Which collection and blueprint should be used for importing products?
     'import' => [
-        'categories' => [
-            'collection' => 'categories',
-            'blueprint' => 'category',
-        ],
         'brands' => [
             'collection' => 'brands',
             'blueprint' => 'brand',
@@ -39,7 +35,7 @@ return [
         'configure' => true,
 
         // The attribute id used for the brand resource.
-        'brand_attribute_id' => 83,
+        'brand_attribute_id' => 139,
 
         // Only show these product visibilities.
         // VISIBILITY_NOT_VISIBLE    = 1;
@@ -58,14 +54,12 @@ return [
 
             \Rapidez\Statamic\Models\Category::class => [
                 'name' => 'Categories',
-                'read_only' => true,
                 'title_field' => 'name',
                 'cp_icon' => 'array',
             ],
 
             \Rapidez\Statamic\Models\Brand::class => [
                 'name' => 'Brands',
-                'read_only' => true,
                 'title_field' => 'value_store',
                 'cp_icon' => 'tags',
                 'order_by' => 'sort_order',
