+++ config/rapidez/statamic/builder.php
@@ -1,5 +1,6 @@
 <?php
 
+use Rapidez\Statamic\Collections\Brands;
 use Rapidez\Statamic\Collections\Categories;
 use Rapidez\Statamic\Collections\Products;
 
@@ -7,5 +8,11 @@ return [
     'collections' => [
         Products::class,
         Categories::class,
+        Brands::class,
+    ],
+    'routes' => [
+        Brands::class => [
+            'default' => '{slug}'
+        ],
     ]
 ];
