--- resources/blueprints/collections/brands/brand.yaml
@@ -1,39 +1,0 @@
-sections:
-  content:
-    display: Content
-    fields:
-      -
-        handle: title
-        field:
-          type: text
-          required: true
-          validate:
-            - required
-          localizable: true
-      -
-        import: content
-        display: 'Description'
-        prefix: 'brand_'
-      -
-        handle: linked_brand
-        field:
-          max_items: 1
-          mode: default
-          resource: brand
-          create: false
-          display: 'Linked brand'
-          type: belongs_to
-          icon: belongs_to
-          listable: hidden
-          instructions_position: above
-          visibility: visible
-          localizable: true
-      -
-        import: image
-      -
-        import: page_builder
-  meta:
-    display: Meta
-    fields:
-      - import: meta_data
-title: brands
