--- resources/blueprints/collections/categories/category.yaml
@@ -1,29 +1,0 @@
-sections:
-  content:
-    display: Content
-    fields:
-      -
-        handle: title
-        field:
-          type: text
-          required: true
-          validate:
-            - required
-          localizable: true
-      -
-        handle: linked_category
-        field:
-          max_items: 1
-          mode: default
-          resource: category
-          create: false
-          display: 'Linked category'
-          type: belongs_to
-          icon: belongs_to
-          listable: hidden
-          instructions_position: above
-          visibility: visible
-          localizable: true
-      -
-        import: page_builder
-title: Categories
