--- resources/blueprints/collections/pages/page.yaml
@@ -1,37 +1,0 @@
-sections:
-  main:
-    display: Main
-    fields:
-      -
-        handle: title
-        field:
-          type: text
-          required: true
-          validate:
-            - required
-          localizable: true
-      -
-        import: page_builder
-  sidebar:
-    display: Sidebar
-    fields:
-      -
-        handle: slug
-        field:
-          type: slug
-          localizable: true
-      -
-        handle: parent
-        field:
-          type: entries
-          collections:
-            - pages
-          max_items: 1
-          listable: false
-          localizable: true
-  meta:
-    display: Meta
-    fields:
-      -
-        import: meta_data
-title: Page
