--- resources/content/collections/brands.yaml
@@ -1,13 +1,0 @@
-title: 'Brand content'
-propagate: true
-template: rapidez-statamic::brands/show
-layout: layout
-revisions: false
-sort_dir: asc
-date_behavior:
-  past: public
-  future: private
-preview_targets:
-  -
-    label: Entry
-    url: '{permalink}'
