--- resources/content/collections/categories.yaml
@@ -1,8 +1,0 @@
-title: 'Category content'
-propagate: true
-revisions: false
-slugs: false
-sort_dir: asc
-date_behavior:
-  past: public
-  future: private
