--- resources/content/collections/pages.yaml
@@ -1,15 +1,0 @@
-title: Pages
-propagate: true
-template: rapidez-statamic::page
-revisions: false
-route: '/{parent_uri}/{slug}'
-sort_dir: asc
-date_behavior:
-  past: public
-  future: private
-preview_targets:
-  -
-    label: Entry
-    url: '{permalink}'
-structure:
-  root: true
