+++ resources/views/brands/show.blade.php
@@ -17,9 +17,9 @@
         @if ($title)
             <h1 class="font-bold text-3xl mb-5">{{ $title }}</h1>
         @endif
-        @if ($brand_content)
+        @if ($content)
             <div class="prose prose-green max-w-none">
-                @include('rapidez-statamic::page_builder.content', ['content' => $brand_content])
+                @include('rapidez-statamic::page_builder.content', ['content' => $content])
             </div>
         @endif
         <x-rapidez::listing query="{ terms: { '{{ $brandAttribute->code }}.keyword': ['{{ $title }}'] } }"/>
