+++ src/Contracts/ImportsBrands.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Rapidez\Statamic\Contracts;
+
+use Rapidez\Statamic\Actions\StatamicEntryAction;
+
+interface ImportsBrands
+{
+    public function handle();
+}
