+++ src/Models/Brand.php
@@ -2,20 +2,27 @@
 
 namespace Rapidez\Statamic\Models;
 
+use Illuminate\Database\Eloquent\Attributes\ObservedBy;
 use StatamicRadPack\Runway\Traits\HasRunwayResource;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Facades\DB;
+use Rapidez\Statamic\Models\Traits\HasContentEntry;
+use Rapidez\Statamic\Observers\RunwayObserver;
 use Statamic\Facades\Site;
 use Statamic\Statamic;
 
+#[ObservedBy([RunwayObserver::class])]
 class Brand extends Model
 {
-    use HasRunwayResource;
+    use HasRunwayResource, HasContentEntry;
 
     protected $table = 'eav_attribute_option';
 
     protected $primaryKey = 'option_id';
+
+    public string $linkField = 'linked_brand';
+    public string  $collection = 'brands';
 
     protected static function booting()
     {
